% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scQCEA.R
\name{CellTypeEnrichmentAnalysis}
\alias{CellTypeEnrichmentAnalysis}
\title{Cell-Type Enrichment Analysis of scRNA-seq data}
\usage{
CellTypeEnrichmentAnalysis(
  project_id,
  gex_library_id,
  input.dir,
  backend.data.dir,
  organism,
  parent.dir
)
}
\arguments{
\item{project_id}{Name of a folder which contains all input files}

\item{gex_library_id}{Name of a folder which contains folders of gene expression matrix for several samples}

\item{backend.data.dir}{Name of a folder which contains reference gene sets for different cell-types}

\item{organism}{hsapiens or mmusculus}
}
\value{
You can find the results as pdf and text files including cell type assignment matrix, tSNE, UMAP, and heatmap plots.
}
\description{
A Cell Type Annotation function for Single-Cell RNA-seq Data.
}
\examples{
project_id='P180121'
gex_library_id=	'481207_03'
input.dir='/10X-gex/481207_03'
backend.data.dir='/Back_End_Data'
organism='hsapiens'
parent.dir='/users/buck/ani054/scQCEA/inst/extdata'
CellTypeEnrichmentAnalysis(project_id, gex_library_id, input.dir, backend.data.dir, organism, parent.dir)
}
\author{
{Isar Nassiri}
}
